add documentation to the internal IR
