+++ cwe_checker_rs/src/intermediate_representation/expression.rs
@@ -2,28 +2,55 @@ use super::ByteSize;
 use super::Variable;
 use crate::prelude::*;
 
+/// An expression is a calculation rule
+/// on how to compute a certain value given some variables (register values) as input.
+/// 
+/// The basic building blocks of expressions are the same as for Ghidra P-Code.
+/// However, expressions can be nested, unlike original P-Code.
+/// 
+/// Computing the value of an expression is a side-effect-free operation.
+/// 
+/// Expressions are typed in the sense that each expression has a `ByteSize`
+/// indicating the size of the result when evaluating the expression.
+/// Some expressions impose restrictions on the sizes of their inputs
+/// for the expression to be well-typed.
+/// 
+/// All operations are defined the same as the corresponding P-Code operation.
+/// Further information about specific operations can be obtained by looking up the P-Code mnemonics in the
+/// [P-Code Reference Manual](https://ghidra.re/courses/languages/html/pcoderef.html).
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Hash, Clone)]
 pub enum Expression {
+    /// A variable representing a register or temporary value of known size.
     Var(Variable),
+    /// A constant value represented by a bitvector.
     Const(Bitvector),
+    /// A binary operation.
+    /// Note that most (but not all) operations require the left hand side (`lhs`)
+    /// and right hand side (`rhs`) to be of equal size.
     BinOp {
         op: BinOpType,
         lhs: Box<Expression>,
         rhs: Box<Expression>,
     },
+    /// A unary operation
     UnOp {
         op: UnOpType,
         arg: Box<Expression>,
     },
+    /// A cast operation for type cast between integer and floating point types of different byte lengths.
     Cast {
         op: CastOpType,
         size: ByteSize,
         arg: Box<Expression>,
     },
+    /// An unknown value but with known size.
+    /// This may be generated for e.g. unsupported assembly instructions.
+    /// Note that computation of an unknown value is still required to be side-effect-free!
     Unknown {
         description: String,
         size: ByteSize,
     },
+    /// Extracting a sub-bitvector from the argument expression.
     Subpiece {
         low_byte: ByteSize,
         size: ByteSize,
@@ -31,6 +58,7 @@ pub enum Expression {
     },
 }
 
+/// The type/mnemonic of a binary operation
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Hash, Clone, Copy)]
 pub enum BinOpType {
     Piece,
@@ -69,6 +97,7 @@ pub enum BinOpType {
     FloatDiv,
 }
 
+/// The type/mnemonic of a typecast
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Hash, Clone, Copy)]
 pub enum CastOpType {
     IntZExt,
@@ -78,6 +107,7 @@ pub enum CastOpType {
     Trunc,
 }
 
+/// The type/mnemonic of an unary operation
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Hash, Clone, Copy)]
 pub enum UnOpType {
     IntNegate,
